Visualize spatial patterns of metagenes within a cell cluster

Visualize spatial patterns of metagenes within a cell cluster

Usage

plotMetagene(d, clufitRes, moduleRes, coord, cluster, background = FALSE, ...)

Arguments

d

The output from runseurat

clufitRes

The output from ctsvg_fit

moduleRes

The output from ctsvg_module

coord

The matrix of spatial locations from cellassign

cluster

A character value indicating which cell cluster to plot

background

Whether to plot all cells (FALSE by default)

...

The parameters in the geom_point(...) function

Value

A ggplot2 plot